% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-proteins.R
\name{extract_ints}
\alias{extract_ints}
\title{extract_ints}
\usage{
extract_ints(ints, mass_fit_params, cut_size = 100)
}
\arguments{
\item{ints}{Vector of intensities for every slice.}

\item{mass_fit_params}{Parameter estimates from the polynomial fit.}

\item{cut_size}{Number of subslices per slice (default is 100)}
}
\description{
Cut every slice into tiny subslices.
}
\examples{
mass_fit_params <- c(2.90230588635708,-0.0990513683718742,0.00309367801794186,-4.05829976900451e-05)
ints <- runif(42,0,1)
extr_ints <- extract_ints(ints, mass_fit_params)
}
