% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-fit-curve.R
\name{filter_and_fit}
\alias{filter_and_fit}
\title{Filter_and_fit}
\usage{
filter_and_fit(pg, low_density_threshold = 10)
}
\arguments{
\item{pg}{ProteinGroups data.frame.}

\item{low_density_threshold}{Remove entries from regions with few data points (default is 10).}
}
\description{
Filter the data and apply a fit on it.
}
\examples{
proteinGroups_path <- system.file("extdata", "Conde_9508_sub.txt", package = "pumbaR")
pg <- load_MQ(proteinGroups_path)
mass_fit <- filter_and_fit(pg, low_density_threshold = 2)
}
