% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-fit-curve.R
\name{filter_repeated_entries}
\alias{filter_repeated_entries}
\title{Filter_repeated_entries}
\usage{
filter_repeated_entries(pg, rep_threshold = 0.3)
}
\arguments{
\item{rep_threshold}{Max percentage of appearance. Default is 0.3.}
}
\description{
Remove proteins which were found in too many slices. Those are very usually
contaminants.
}
