% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-fit-curve.R
\name{filter_weak_intensity}
\alias{filter_weak_intensity}
\title{Filter weak intensitites}
\usage{
filter_weak_intensity(pg, int_threshold_percent = 0.5)
}
\arguments{
\item{rep_threshold}{Max percentage of appearance. Default is 0.3.}
}
\description{
Remove entries with a weak intensity.
}
