% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{get_correction_factor}
\alias{get_correction_factor}
\title{Get correction factor}
\usage{
get_correction_factor(ints)
}
\arguments{
\item{ints}{The slice intensities from a proteinGroups.txt file.}
}
\description{
Gives back the total sum of intensities over all slices and proteins.
This correction factor is used for intensity normalisation.
}
\examples{
proteinGroups_path <-
system.file("extdata", "Conde_9508_sub.txt", package = "pumbaR")
pg <- load_MQ(proteinGroups_path)
ints <- get_intensities(pg)
corr_factor <- get_correction_factor(ints)
}
