% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-and-filter.R
\name{get_intensities}
\alias{get_intensities}
\title{Get intensities}
\usage{
get_intensities(pg)
}
\arguments{
\item{pg}{ProteinGroups data.frame.}
}
\description{
Extract the Intensity.H values and put slice numbers as the column names.
}
\examples{
proteinGroups_path <- "/Users/admin/Work/PAF/projects/SliceSILAC/latest/data/Conde_9508/proteinGroups.txt"
pg <- load_MQ(proteinGroups_path)
ints <- get_intensities(pg)
}
