% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-and-filter.R
\name{get_intensity_columns}
\alias{get_intensity_columns}
\title{Get intensitie columns}
\usage{
get_intensity_columns(pg, sample_name = NULL)
}
\arguments{
\item{pg}{ProteinGroups data.frame.}
}
\description{
Extract the Intensity.H or Intensity. columns
}
