% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{get_max_norm_int}
\alias{get_max_norm_int}
\title{Get max normalized intensity}
\usage{
get_max_norm_int(norm_pg)
}
\arguments{
\item{norm_pg}{normalized ProteinGroups data.frame.}
}
\description{
Gives back the maximum normalized intensity.
}
\examples{
proteinGroups_path <-
system.file("extdata", "Conde_9508_sub.txt", package = "pumbaR")
pg <- load_MQ(proteinGroups_path)
norm_pg <- get_normalized_table(pg)
max_norm_int <- get_max_norm_int(norm_pg)
}
