% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{get_normalized_table}
\alias{get_normalized_table}
\title{Get normalize table}
\usage{
get_normalized_table(pg, sample_name = NULL)
}
\arguments{
\item{pg}{ProteinGroups data.frame.}
}
\description{
Gives back the normalized intensities together with some informations
about the proteins (proteinAC, gene name, etc.).
}
\examples{
proteinGroups_path <-
system.file("extdata", "Conde_9508_sub.txt", package = "pumbaR")
pg <- load_MQ(proteinGroups_path)
normalized_table <- get_normalized_table(pg)
}
