% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-and-filter.R
\name{get_slice_numbers}
\alias{get_slice_numbers}
\title{Get the slice numbers}
\usage{
get_slice_numbers(col_names)
}
\arguments{
\item{col_names}{Column names.}
}
\description{
Extract the slice numbers from the column names
}
