% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_intensities}
\alias{normalize_intensities}
\title{Normalize intensities}
\usage{
normalize_intensities(ints)
}
\arguments{
\item{ints}{The slice intensities from a proteinGroups.txt file.}
}
\description{
Normalize intensities by dividing every entry by the total sum of
intensities over all slices and proteins. Gives back a data-frame
with normalized intensities.
}
\examples{
proteinGroups_path <-
system.file("extdata", "Conde_9508_sub.txt", package = "pumbaR")
pg <- load_MQ(proteinGroups_path)
ints <- get_intensities(pg)
norm_ints <- normalize_intensities(ints)
}
