% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-plots.R
\name{plot_fit}
\alias{plot_fit}
\title{plot_fit}
\usage{
plot_fit(pg, mass_fit)
}
\arguments{
\item{pg}{ProteinGroups data.frame.}

\item{mass_fit}{Mass-slice fitting function}
}
\description{
plot the MaxQuant data with the fitting curve
}
