% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-plots.R
\name{plot_MQ}
\alias{plot_MQ}
\title{plot_MQ}
\usage{
plot_MQ(pg)
}
\arguments{
\item{pg}{ProteinGroups data.frame.}
}
\description{
plot the intensities
}
\examples{
proteinGroups_path <- proteinGroups_path <- system.file("extdata", "Conde_9508_sub.txt", package = "pumbaR")
pg <- load_MQ(proteinGroups_path)
print(plot_MQ(pg))
}
