% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-proteins.R
\name{sum_slides}
\alias{sum_slides}
\title{sum_slides}
\usage{
sum_slides(extracted_data)
}
\arguments{
\item{extracted_data}{A list of data extracted by extract_ints}
}
\description{
Sum up slides which were generated by extract_ints
}
