% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-fit-curve.R
\name{filter_and_fit}
\alias{filter_and_fit}
\title{Filter_and_fit}
\usage{
filter_and_fit(pg)
}
\arguments{
\item{pg}{ProteinGroups data.frame.}
}
\description{
Filter the data and apply a fit on it.
}
\examples{
proteinGroups_path <- "/Users/admin/Work/PAF/projects/SliceSILAC/latest/data/Conde_9508/proteinGroups.txt"
pg <- load_MQ(proteinGroups_path)
mass_fit <- filter_and_fit(pg)
plot_fit(pg, mass_fit)
}
