% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-plots.R
\name{plot_MQ}
\alias{plot_MQ}
\title{plot_MQ}
\usage{
plot_MQ(pg)
}
\arguments{
\item{pg}{ProteinGroups data.frame.}
}
\description{
plot the intensities
}
\examples{
proteinGroups_path <- "/Users/admin/Work/PAF/projects/SliceSILAC/latest/data/Conde_9508/proteinGroups.txt"
proteinGroups_path <- "tests/testthat/Conde_9508_sub.txt"
pg <- load_MQ(proteinGroups_path)
print(plot_MQ(pg))
}
